refactor: 패키지 생성 및 클래스 이동
역할에 따라 패키지를 생성, 분리하고 기존 작성 클래스 이동
