+++ src/main/java/vendingmachine/model/Coin.java
@@ -1,4 +1,4 @@
-package vendingmachine;
+package vendingmachine.model;
 
 import java.util.Arrays;
 import java.util.List;
