+++ src/main/java/vendingmachine/util/Validator.java
@@ -1,4 +1,4 @@
-package vendingmachine;
+package vendingmachine.util;
 
 import vendingmachine.message.Error;
